mode blockchain info stuff to its own file
